8251893: jextract does not generate source for ConstantHelper even when --source is used
